+++ tests/test_shell.py
@@ -10,11 +10,18 @@ class Shell_Python(lb.ShellBackend):
 
     binary_path = attr.value.str('python', sets=False)
     path: str = attr.value.Path(None, key=None, must_exist=True, help='path to a python script file')
-    command: str = attr.value.str(None, help='execute a python command')
+    command: str = attr.value.str(None, key='-c', help='execute a python command')
 
-    def run(self, **kwargs):
-        return super().run(
-            self.FLAGS,
+    def get_flags(self):
+        pass
+
+    def get_arg_list(self):
+        [for name in attr._bases.list_value_attrs(self)]
+
+    def __call__(self, *arg_list, **flags):
+
+        return self.run(
+            FLAGS,
             **kwargs
         )
 
@@ -23,7 +30,7 @@ def test_python_print():
     python = Shell_Python()
 
     python.command = f'print("{TEST_STRING}")'
-    assert python.run() == f'{TEST_STRING}\n'.encode()
+    assert python() == f'{TEST_STRING}\n'.encode()
 
     python.command = f'import sys; print("{TEST_STRING}", file=sys.stderr)'
     assert python.run() == ''.encode()
@@ -32,16 +39,3 @@ def test_python_print():
     with pytest.raises(ChildProcessError):
         python.run(check_stderr=True)
 
-
-if __name__ == '__main__':
-    import inspect
-
-    lb.show_messages('debug')
-
-    python = Shell_Python(command='print("hello world")')
-    print(python.run())
-    # print(python['resource'].key is lb.Undefined)
-    # print(lb.Unicode.__defaults__)
-    # print(inspect.signature(lb.Unicode.__init__))
-
-    # print(python._commandline())